Complex API (#302)
Allow the user to discretize complex `SesquilinearForm`, `LinearForm`
and `Functional` and to assemble them. For this, the user needs to add
to the `SymbolicSpace` of SymPDE an attribute named
`codomain_type='complex'`.

Also, add as a comment the case where the user gives a
`SesquilinearForm` (not yet implemented in SymPDE), which will be
discretized into an object of type `DiscreteSesquilinearForm`.

The different projectors can now support complex values.

Adds some tests for the complex assembly in `test_assembly.py`, some
tests to check the precision of the solution in 'test_complex.py' for a
domain with and without mapping, both single- and multi-patch.

Add `domain_dtype` property to `DefNode`, to build the correct decorator
for constant in Pyccel. Also, this parameter corrects the importation of
mathematical function in the complex case from `math` to `numpy`, thus
the function can handle the complex values.

NOTE: from now on when declaring a SymPDE `Constant` for Psydac, the
type of the constant must be specified with one of the boolean flags
`complex`, `float`, or `integer`.

Add `dtype`property to :
- MatrixGlobalBasis
- MatrixLocalBasis
- MatrixQuadrature
- BlockStencilMatrixLocalBasis
- BlockStencilMatrixGlobalBasis
- BlockStencilVectorLocalBasis
- BlockScalarLocalBasis
- StencilMatrixLocalBasis
- StencilVectorLocalBasis
- StencilVectorGlobalBasis
- EvalField
- LocalElementBasis
- GlobalElementBasis
- TensorIterator
- MatrixRankFromCoords
- LocalElementBasis
- ScalarLocalBasis
- ComputeLogicalBasis
- GeometryExpr
- GeometryExpressions
- Zeros

Add the `prefix` property to : 
- BasisAtom
- IteratorBase
- LogicalBasisValue
- Loop

Add `change_dtype` function, which copies the `vector_space`of a
`FemSpace` and changes its data type.

---------

+++ psydac/fem/splines.py
@@ -166,7 +166,7 @@ class SplineSpace( FemSpace ):
         return self._histopolation_grid
 
     # ...
-    def init_interpolation( self ):
+    def init_interpolation( self, dtype=float ):
         """
         Compute the 1D collocation matrix and factorize it, in preparation
         for the calculation of a spline interpolant given the values at the
@@ -190,7 +190,7 @@ class SplineSpace( FemSpace ):
             l = abs( dmat.offsets.min() )
             u =      dmat.offsets.max()
             cmat = csr_matrix( dmat )
-            bmat = np.zeros( (1+u+2*l, cmat.shape[1]) )
+            bmat = np.zeros( (1+u+2*l, cmat.shape[1]), dtype=dtype )
             for i,j in zip( *cmat.nonzero() ):
                 bmat[u+l+i-j,j] = cmat[i,j]
             self._interpolator = BandedSolver( u, l, bmat )
@@ -200,7 +200,7 @@ class SplineSpace( FemSpace ):
         self._interpolation_ready = True
 
     # ...
-    def init_histopolation( self ):
+    def init_histopolation( self, dtype=float):
         """
         Compute the 1D histopolation matrix and factorize it, in preparation
         for the calculation of a spline interpolant given the integrals within
@@ -224,7 +224,7 @@ class SplineSpace( FemSpace ):
             l = abs( dmat.offsets.min() )
             u =      dmat.offsets.max()
             cmat = csr_matrix( dmat )
-            bmat = np.zeros( (1+u+2*l, cmat.shape[1]) )
+            bmat = np.zeros( (1+u+2*l, cmat.shape[1]), dtype=dtype)
             for i,j in zip( *cmat.nonzero() ):
                 bmat[u+l+i-j,j] = cmat[i,j]
             self._histopolator = BandedSolver( u, l, bmat )
